+++ src/test/java/de/dm/intellij/liferay/language/resourcebundle/LiferayResourceBundlePropertiesImplicitUsageProviderTest.java
@@ -29,7 +29,7 @@ public class LiferayResourceBundlePropertiesImplicitUsageProviderTest extends Ba
     public void testImplicitUsageArbitraryPropertyInLanguageProperties() {
         myFixture.configureByFiles("Language_unused.properties");
 
-        //Language.properties should not show any unused warning, even if javax.portlet.title.my_portlet is not used explicitly
+        //Language.properties should show unused warning for other properties
         myFixture.checkHighlighting();
     }